+++ frontend/language/type.ts
@@ -1,3 +1,3 @@
-import * as nb from './nb.json';
+import * as nb from './src/nb.json';
 
 export type TranslationKey = keyof typeof nb;
